Fixes https://github.com/tkoolen/RigidBodyDynamics.jl/issues/93 (in an unsatisfactory way).

'RTTI'-style dispatch inspired by https://groups.google.com/d/msg/julia-users/ude2-MUiFLM/z-MuQ9nhAAAJ, hopefully a short-term solution.

By specifying the exact set of types for which the Joint methods work, the compiler is able to figure out that the return type is always the same, so the type instability goes away.

Unfortunately this means that it is currently no longer possible for a user to add a new joint type, which needs to be fixed at some point.

Before (on laptop):
```
  "inverse_dynamics" => BenchmarkTools.Trial: 
  samples:          10000
  evals/sample:     1
  time tolerance:   5.00%
  memory tolerance: 1.00%
  memory estimate:  60.86 kb
  allocs estimate:  869
  minimum time:     115.93 μs (0.00% GC)
  median time:      124.37 μs (0.00% GC)
  mean time:        137.13 μs (4.97% GC)
  maximum time:     3.22 ms (92.80% GC)
  "dynamics" => BenchmarkTools.Trial: 
  samples:          10000
  evals/sample:     1
  time tolerance:   5.00%
  memory tolerance: 1.00%
  memory estimate:  75.22 kb
  allocs estimate:  1116
  minimum time:     214.63 μs (0.00% GC)
  median time:      221.07 μs (0.00% GC)
  mean time:        251.15 μs (3.65% GC)
  maximum time:     5.27 ms (90.13% GC)
  "mass_matrix" => BenchmarkTools.Trial: 
  samples:          10000
  evals/sample:     1
  time tolerance:   5.00%
  memory tolerance: 1.00%
  memory estimate:  57.94 kb
  allocs estimate:  674
  minimum time:     95.19 μs (0.00% GC)
  median time:      101.76 μs (0.00% GC)
  mean time:        111.73 μs (4.83% GC)
  maximum time:     3.35 ms (94.30% GC)
```
After:
```
  "inverse_dynamics" => BenchmarkTools.Trial: 
  samples:          10000
  evals/sample:     1
  time tolerance:   5.00%
  memory tolerance: 1.00%
  memory estimate:  35.19 kb
  allocs estimate:  466
  minimum time:     103.59 μs (0.00% GC)
  median time:      110.43 μs (0.00% GC)
  mean time:        118.68 μs (2.83% GC)
  maximum time:     3.64 ms (93.65% GC)
  "dynamics" => BenchmarkTools.Trial: 
  samples:          10000
  evals/sample:     1
  time tolerance:   5.00%
  memory tolerance: 1.00%
  memory estimate:  49.55 kb
  allocs estimate:  713
  minimum time:     203.39 μs (0.00% GC)
  median time:      207.81 μs (0.00% GC)
  mean time:        232.11 μs (2.54% GC)
  maximum time:     3.91 ms (84.58% GC)
  "mass_matrix" => BenchmarkTools.Trial: 
  samples:          10000
  evals/sample:     1
  time tolerance:   5.00%
  memory tolerance: 1.00%
  memory estimate:  40.98 kb
  allocs estimate:  488
  minimum time:     89.71 μs (0.00% GC)
  median time:      95.81 μs (0.00% GC)
  mean time:        103.78 μs (2.98% GC)
  maximum time:     2.38 ms (91.24% GC)
```